+++ src/providers/arcgisrest/qgsafsshareddata.cpp
@@ -37,7 +37,7 @@ bool QgsAfsSharedData::getFeature( QgsFeatureId id, QgsFeature &f, const QgsRect
   }
 
   // Fetch 100 features at the time
-  int startId = ( id / 100 ) * 100;
+  int startId = static_cast<int>( ( id / 100 ) * 100 );
   int stopId = std::min( startId + 100, mObjectIds.length() );
   QList<quint32> objectIds;
   objectIds.reserve( stopId );
